requirements.txt
openai-whisper @ git+https://github.com/openai/whisper.git